Initial Checkin Of the Atomic Integers
This is the first cut of the atomic integers in DKit. It's the first
component of DKit as well, and while I'd like to have time to write more
tests, I'm pretty confident that we're in good shape so far. This all
builds on Mac OS X 10.7 and seems to work just as well on Linux.