Switched to Volatile and Added Inequalities
I wanted to make sure that the compiler knew that the ivars are volatile and
not count on them being stable on a single core. Also, while I was working on
these guys, I added in the remainder of the simple inequalities for <, <=,
>, >= with the basic datatypes as well as the other members of this group.
This makes it much easier to deal with unsigned values compared to signed, as
I take care of the signed/unsigned in the operator code. Simple. Clean.