Fixed Up Problems in CircularFIFO
I had a few bugs in the FIFO code - like the wrong size calculation and the
wrong capacity value. I was working to get something out, and I didn't take
enough time to verify the code until I wrote the tests - that's when I saw
that the size was wrong. Additionally, the initial push() was wrong and that
needed to be fixed as well. In all, a good reason to have the tests. :)