Added the Util Timer Class
In order to have good tests, we needed to have a good timer class. This was
a little tricky because the preferred method is really to use the
clock_gettime() function, but that doesn't exist on Mac OS X, and so I had
to find a Mac OS X version of the code to ifdef in there. This was a very
interesting find, and it makes the timer all that more valuable to me.