Updated the README for the FIFO and Timer
With the addition of the CircularFIFO and the timer classes, it made sense to
update the README so that they are documented here up-front. Nothing fancy,
and I want to expand on the timer class a lot, but for now, it's nice to be
able to get something down for these initial cases.