Added Non-Referenced Timer Method
I found that there's a far faster version of the nanosecond timer in Mac OS X
but it's refrence-less, so I added in a simple second timer that's great for
getting just a refrence-less timestamp. I ifdef-ed the Mac OS X part out, and
used the clock_gettime() for the linux side as it's pretty darn fast as well.
In all, it's a nice addition to the timer class.