Fixed the Fixed-Point Math in the Timer
In the timer, the 'factor' was going to be zero because it was a fraction and
for an integer that's zero. It made sense to just leave it as-is in the code
and then divide the system call by 1000 to get it to usec right off the bat.
This works just fine now.