Updated the Docs for the Non-Reference Timer
With the addition of the more efficient Mac OS X non-reference timer, I wanted
to update the docs so that they discussed the addition and the reason for it.
There will be lots of times that there's no need to spend the effort to get
a referenced time as it's just an interval timer.