Fixed up the Comment on the Method
It wasn't true that we were returning a 'reference' to the top of the queue -
we are returnign a COPY of the top of the queue. That's important in those
cases where we are dealing with complex objects, but when it's basic data
types and pointers, it's less of an issue. Still... it's important to be
right.