Created the MP/SC Linked FIFO Queue
This template class is the multi-producer, single-consumer, linked FIFO queue,
so named because it uses a singly-linked list to hold all the elements. This
has the downside of needing to allocate storage for each push(), but the
benefit of not having to worry about pre-allocating a large circular FIFO and
hitting it's limits. This is far more flexible, and uses only what's needed,
but it comes at a cost.