Created Test for MP/SC LinkedFIFO
I wanted to be able to make sure that the LinkedFIFO was working, and at the
same time, I wanted to have some sense of relative speed with respect to the
other queues in the package. It's clear that we're paying a proce for this
flexibility, but it's still nice to have when that flexibility is important.