Fixed the Loop Type and CAS on Head
There are two small fixes here - the correct use of size_t as opposed to
uint32_t in the loop, and the swap out of the '++_head' with a nice CAS op
to do the same thing. This is just to make sure that there's no chance that
the _head is compromised in any way.