Added MP/SC CircularFIFO
We needed to have a MP/SC circular FIFO container, and now we do. This is
very much like the SP/SC circular FIFO, but it uses a slightly different
data structure in order to make it possible to have multiple producers. It's
simple, and clean, and should work quite well.