Added Test for MP/SC CircularFIFO Class
With the addition of the MP/SC CircularFIFO, it's important that we have a
suite of tests to make sure that things are functioning as they should. One of
the things I was most concerned about, and remains to be tested, is the case
where you have multiple threads all crashing the queue at once and trying to
back-out their crashes effectively. I believe the logic is sound, but the
tests still need to be done in order to confirm it.