Added a SP/MC Linked FIFO to the Library
I was thinking that I needed to have a SP/MC FIFO in order to make a simple
load balancer and other job-distribution schemes. So I looked at what I needed
to do in order to get that working, and this is the first cut. I need to do
more tests on it to make sure it's completely safe, but it appears to be.