Fixed the Nearly Empty push() Problem
We had a problem with the push() code when we were nearly empty. It was
possible to have a race condition where the last node was being stripped and
used in the pop() method at the same time we're trying to add in a new node
in the push() method, and the reference to '_tail->next' would possibly become
undefined. What I needed to do was to make sure that there's only one case
I wanted, and then use a simple CAS operation on it to make sure that we got
the appropriate behavior. This should work much better.