Added SP/MC CircularFIFO to Package
After looking at the existing MP/SC CircularFIFO, I realized that if I did
just a little work, I could make a somewhat less efficient SP/MC version
and so I decided to give that a try. This has a very similar structure to
the MP/SC, but it's in the use of the head and tail that's really different.
We didn't want to use them for size, as that was a potential problem, so I
just added a new ivar and updated that directly. Now I should be OK to go.