Added FIFO<T> SuperClass to Queues
In the midst of working on the multi-threaded tests for the queues, I realized
that it would be a LOT better if there were one (abstract) base class for all
the FIFO implementations - so I created it. This allows us to more easily
use these in code as we only have to deal with the instantiation and after
that, it's a simple base class usage to get all the functionality we need.
This is going to make it a lot cleaner to move from one instance to another.