Added Multi-Threaded Tests to LinkedFIFO
We needed to be able to test the LinkedFIFOs properly, and that means being
able to hammer them with producers and consumers. In order to do this, we
needed to make a Hammer and a Drain to help out. Then I had to add in the
code on the test app to make the appropriate numbers of these, and then use
them to exercise the different LinkedFIFO implementations we had. Because
we are now based on the FIFO<T> base class, the implementation of the
Hammer and Drain weren't bad at all.