Added 'tests' Target to Makefile
Now that we have a few tests, it makes sense to have a way to run all of them
one right after another. This is just a simpler way than having a script as
it's all in the Makefile. It will be something to look at unit testing
frameworks for Boost and others to see if we can make something a little more
uniform than the apps I've been writing.