Added Source and Sink Template Classes
The next step in DKit is the inclusion of a Source/Sink scheme so that you
can start wiring these lockless queues together in a pub/sub like fashion
with very little effort. The Source and Sink are template classes so it's
very easy to pick what data they pass around, and therefore it's easy to
make them very powerful. I've used something like these in data feeds and
high through-put servers, and they are very useful.