Added in Timestamp Formatting Calls
There are plenty of times that we would like to see the timestamps as formatted
strings and not just uint64_t values in microseconds. For this reason, we added
several formatting methods to the timer class. They can format the entire
timestamp (date and time), or just the date, or just the time. For the times,
you can choose to see the microseconds, or just to the nearest second. It's
pretty decent, but if we need more, we can always add that later.