Fixed Buffer Overrun Bug
I had a problem with the way the SP/SC Circular FIFO was working when it got
to the wrap-around case. I wasn't properly dealing with the _head and _tail
in the same "segment" of the wrap. What I needed to do was to simplify things
quite a bit, and realize that I'm not going to be able to atomically update
the head. I needed to add and then mask it, and that's two operations, and
so I just went ahead and used the simpler code. It's still logically fine,
but it's not as "clever", but that was the bug, wasn't it? :)