Added Pools to DKit
This is really exciting to me. In the past, I had completely different
versions for pools of strings, or datagrams, or whatever, and with this
code, I've been able to successfully templatize it so that I can have ONE
pool class and with it define not only the data to hold and the size, but
also the access scheme of producers and consumers. This means that I really
can make about any pool I need and not need to worry about what it's
access methodology is. Very nice work.