Cleaned Up Whitespace and Makefile Dependencies
I didn't like the lack of a space in the template function definitions, so I
added it there to match the same space in the class definition. No functional
change otherwise. I also realized that the makefile dependencies were for .o
files, and there are none of those in the test suite, so I changed the suffix
argument to make depend and now it's properly creating the entries for the
test apps themselves. Much better.