Converted to Consistent ivar Naming Scheme
I had three classes that used my old ivar naming scheme where the 'm' prefix
on the variable meant "member", and there's really no need for that. The
underscore is sufficient as they are small classes and there are no static
ivars to have to deal with. It's just a way to make sure that all the classes
have a consistent coding standard.