Added UDP Receiver Test App
With the inclusion of the new udp_receiver class, I needed to have a simple
way to test it, so I made one. It's nothing fancy, but it points out that
if you use the boost ASIO sender to send it data, it'll detect that things
are all done, and shutdown gracefully. It also shows how the io_service
sharing works. Very nice.