Added I/O Section to DKit - UDP Receiver
The next extension to DKit is the I/O package - that which I've wanted to
build since the beginning of the DKit work. The idea is simple - make use
of the lockless containers, atomics, source, sink, and pool, and use them
to make a UDP listener that is highly efficient, but exceptionally
customizable for the user. The UDP receiver delivers datagrams to the
registered listers (sink<datagram*> subclasses), and it's up to the
listeners to do with those datagrams as they see fit. This is the entry
point for an exchange feed system where these datagrams will be converted
into messages, and sent to waiting clients.