Updated the Docs for the I/O Package
Now that I've added in the UDP receiver, I needed to make it clear what the
reasons for adding it were, and why it's built the way it is. I also decided
that it was a good idea to put in just a little example of how you might
use it.