Added Protected Accessor Methods
When subclassing the sink and source, it became clear that I needed to have
SOME way of getting to the sets of registered sinks and sources of each side.
What made MOST sense was simply making a protected accessor method which could
be called from the subclass and used as if it were the data. This didn't
require us to expose the data or make difficult 'friend' relationships.