Created Adapter Template Class
Once I had the source and sink classes working, it was clear that the next
thing we'd need was an 'adapter' class - something that takes one kind of
data and generates another. Maybe these are the same, maybe not. But this
new template class doesn't care. It's subclassing the source and sink and
using them to make the necessary connections and registrations, so that we
don't copy-n-past code that we don't need to. This allows us to make the
exchange codecs as simple as subclassing the adapter, taking a datagram and
generating a message (of some kind). Very nice.