Added Adapter Test to UDP Datagram Tests
I wanted to be sure that the adapter was working, so I added it as a simple
part of the UDP datagram test app. It's working just like it should, and
while the implementation of the adapter is very simple, it could clearly be
far more complex, and include buffers, FIFO queues, and threads for processing.