Updated Docs to Add in the Adapter Class
Now that I added in the adapter class, I needed to update the docs to make it
clear what it was, why we made it, and how it's related to the source and
sink. Pretty simple stuff - to use, but it was interesting getting it all
to work out easily. Templates can be tricky.