Added #ifdef on the Thread Pooling Definitions
Because we're going to use the same thread pooling on the UDP transmitter as
the UDP receiver, it makes sense to #ifdef the typedefs so that we don't
get ourselves into trouble with double definitions of these things. It's a
simple step to make sure that we don't do this, and this is the receiver's
part.