Created and UDP Transmitter
We needed to have a UDP transmitter subclassed off the sink<T> to round-out
the initial I/O package of DKit. This will take <datagram*> instances and
will send them out (async) on the provided UDP multicast channel. This is
quite a bit simpler than the UDP receiver, but that's the nature of the
send versus receive for sockets. We also included this in the test app to
show how you'd use it as a repeater of UDP data. Simple. Effective.