Fixed Incomplete Implementation of Transmitter
I had forgotten to complete the onMessage() call, and while it was simple, it
was ESSENTIAL that I call the asyncSend() method to get the datagrams out on
the wire. Silly of me to have forgotten it. Also, I needed to finish the
socket settings, and the constructors needed to call the init() method when
given all the parameters as the caller is likely to believe that these are
ready to go, and if they aren't they just will appear to "not work". So I added
in the calls to init_nl() in the constructors where we had sufficient data to
initialize the sockets.