Little Fixes for Compiling on Ubuntu 12.04
I pulled the code to a new Ubuntu 12.04 machine and worked on the compile
until I got all the things fixed that needed to be fixed. This included a
bunch of includes that needed to be defined, as well as clearing out some
unused code. The biggest thing was the constructors on the thread_info in
the udp_receiver and udp_transmitter - I didn't have them and the default
constructor and copy-constructor were fine on Mac OS X, but on Ubuntu 12.04,
they needed to be defined. Specifically, the problem with the copy-constructor
for the boost::thread. Now that I have this all fixed, it's compiling and
working fine for 64-bit builds on Ubuntu 12.04.