Added Trie to DKit
I wanted to add a simple trie, but what I was able to put together is far
beyond what I'd initially hoped. This version has a variable key size, as
well as handling the memory management for pointers as well as plain old
datatypes. Additionally, we're using functors on the Nodes to enable the
caller to do just about anything they want to the data, and not have to
worry about building it into the trie. This is a really great version of
the container, and it's something I'm really glad we have in DKit now.