Cleaned up clear() on Node, and Added remove()
I needed to be a little more careful on the Node::clear() method as I needed
to CAS out the value if it was a pointer. A simple assignment was unwise. So
with that fixed, I decided to add clear() and remove() methods to the trie
so that we can clear out the Nodes as well as extract the values. This will
become clearer when I make the next component - the conflation queue.