Cleaned Up Comments and Protected enum
I needed to protect the definition of the enum for the type of the queues
because I'm about to include it in another template class, and I didn't want
it to be multiply defined. I also saw that there was a silly typo in the
comment about the final #endif, so I cleaned that up too.