Disallowed the Assignment Operator
One of the problem we face with the trie is the assignment operator. The
duplication of PODs is simple -- an assignment. But the duplicateion of a
pointer is not nearly as easy. We can't do a shallow copy or we run the risk
of double-freeing all the pointers. So until we enforce a clone() function
or something that will cleanly duplicate the pointers without leaks or double
frees, we need to just disallow it.