Added the upsert() Method to Insert/Update Value
When placing the value in the trie, one thing I needed was to be able to add
it and know if it's a new value, or replacing an old one. The reason being
that the conflation queue is going to need this, and while it might only be
needed there, it's essential, so it's in.