Added Conflation Queue
This is the conflation queue that looks at the key_value() of each item
as it's push()-ed onto the queue, and replaces any old values on the queue
without altering the item's place in the queue. This is really quite useful
in market data processing when you need the latest data, and don't have time
to process all the data. It's made from the other components of DKit, and
initial tests show it to be working just fine.