Fixed Problem of POD vs. Pointers in Trie
I didn't check the handling of the POD vs. pointers in the trie - I was always
testing with pointers, so I needed to go back into the code and make sure that
I was properly handling the two different styles of assignment, etc. Turns out
I wasn't. But the fix was sitting in the pool<>... I just needed to bring in
the concept of the template function destroy() and then make one for POD and
one for pointers, and then I was golden. Works like a charm.