Fixed up the code for macOS 10.13
We had a few issues that needed to be cleared up for macOS 10.13, and so
these are they. It's not a lot, and a few of these are really things
that I should have fixed up in the past, but the older C++ compiler
allowed me to get away with. All the tests look good, so we're off to
the races.