Added the TCP endpoints
We needed to be able to deal with the TCP receiver and transmitter for
some potential work, and so it made sense to add it in. This is just a
simple conversion of the UDP versions, but with enough changes to make
it work with the reliable TCP sockets, and the difference in Boost's
socket implementation. Not too bad at all.